requirements.txt
numpy>=1.26
pandas>=2.2
Pillow>=10.0
pyyaml>=6.0
scikit-learn>=1.5
tqdm>=4.66
